backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# AI Provider SDKs
anthropic==0.40.0
google-generativeai==0.8.3
openai==1.58.1

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0

# Database
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
asyncpg==0.30.0

# Authentication (Clerk JWT verification)
PyJWT==2.10.1
cryptography==44.0.0

# Payments (Stripe)
stripe==11.4.1

# Document Processing (for future phases)
python-docx==1.1.2
markdown==3.7

# Security
slowapi>=0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
